Minor improvements and md formatting tweaking
Not loading fzf since I'm going to mainly use Vim's built in file navigator () and NerdTree for now. Still working on markdown header formatting. The plasticboy plugin groups it as html syntax in local md files. But even the code I added doesn't seem to work. It only works with github sourced md files, which is likely due to the git-fugitive plugin interacting with the syntax. And even then the headers are not bolded (though they are a different color). And to nitpick even further, the header hashes aren't concealed, but that could be something I need to enable. Who knew it would be this difficult to change a markdown header...