Fixing custom markdown headers, and adding some leader key shortcuts
Finally figured out how to override the plasticboy plugin for markdown header styling. Currently the entire markdown header, including the delimiter, is set to my custom color. I had been looking for a way to conceal the delimiters in md headings, but couldn't figure it out. Giving up on that for now, though there may be a way to enable or override some default vim or plasticboy setting to that end. Also the whole header is automatically styled in bold so it stands out.